@@ -1,27 +1,17 @@
 ﻿using QPlayer.Models;
-using Cue = QPlayer.Models.Cue;
+using QPlayer.SourceGenerator;
+using QPlayer.ThemesV2;
+using QPlayer.Views;
 
-namespace QPlayer.ViewModels
+namespace QPlayer.ViewModels;
+
+[Model(typeof(GroupCue))]
+[View(typeof(CueEditor))]
+[DisplayName("Group Cue")]
+[Icon("IconGroupCue", typeof(Icons))]
+public class GroupCueViewModel : CueViewModel
 {
-    public class GroupCueViewModel : CueViewModel, IConvertibleModel<Cue, CueViewModel>
+    public GroupCueViewModel(MainViewModel mainViewModel) : base(mainViewModel)
     {
-        public GroupCueViewModel(MainViewModel mainViewModel) : base(mainViewModel)
-        {
-        }
-
-        public override void ToModel(Cue cue)
-        {
-            base.ToModel(cue);
-        }
-
-        public static new CueViewModel FromModel(Cue cue, MainViewModel mainViewModel)
-        {
-            GroupCueViewModel vm = new(mainViewModel);
-            if (cue is GroupCue gcue)
-            {
-                //
-            }
-            return vm;
-        }
     }
 }
