@@ -1,5 +1,5 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
-using ReactiveUI.Fody.Helpers;
+using QPlayer.SourceGenerator;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,9 +9,9 @@ using System.Windows;
 
 namespace QPlayer.ViewModels;
 
-public class ProgressBoxViewModel : ObservableObject
+public partial class ProgressBoxViewModel : ObservableObject
 {
-    [Reactive] public Visibility Visible { get; set; } = Visibility.Collapsed;
-    [Reactive] public string Message { get; set; } = "Loading...";
-    [Reactive] public float Progress { get; set; }
+    [Reactive] private Visibility visible = Visibility.Collapsed;
+    [Reactive] private string message = "Loading...";
+    [Reactive] private float progress;
 }
